+++ test/int/test_http_server.py
@@ -24,3 +24,12 @@ def test_server_works(cluster: Cluster):
     instance.wait_online()
     with urlopen("http://127.0.0.1:8080/hello") as response:
         assert b"world" == response.read()
+
+
+def test_webui(cluster: Cluster):
+    instance = cluster.add_instance(wait_online=False)
+    instance.env["PICODATA_HTTP_LISTEN"] = ":8081"
+    instance.start()
+    instance.wait_online()
+    with urlopen("http://127.0.0.1:8081") as response:
+        assert "text/html" == response.headers.get("content-type")